refactor: assignment compliance updates
