+++ _P1/Logic/src/main/java/vdm/p1/logic/objects/base/GoToSceneButton.java
@@ -19,8 +19,7 @@ public abstract class GoToSceneButton extends Button {
 
 	@Override
 	public boolean onPrimaryAction(TouchEvent event) {
-		Logic logic = (Logic) engine.getLogic();
-		logic.changeScene(createScene());
+		engine.getLogic().setScene(createScene());
 		return true;
 	}
