+++ _P1/PCEngine/src/main/java/vdm/p1/pcengine/DesktopGraphics.java
@@ -163,7 +163,6 @@ public final class DesktopGraphics implements IGraphics {
 
 	@Override
 	public void save() {
-
 	}
 
 	@Override
