+++ _P1/PCLauncher/src/main/java/vdm/p1/pclauncher/Program.java
@@ -7,6 +7,7 @@ import vdm.p1.pcengine.DesktopEngine;
 public class Program {
 	public static void main(String[] args) {
 		DesktopEngine engine = new DesktopEngine();
+		engine.getGraphics().setResolution(400, 600);
 		Logic logic = new Logic(engine);
 		engine.setLogic(logic);
 		engine.run();
